+++ crates/firehose/src/runner.rs
@@ -8,7 +8,7 @@ use reth_ethereum_forks::EthereumHardforks;
 use reth_evm::execute::BlockExecutor;
 use reth_exex::{ExExContext, ExExEvent};
 use reth_provider::{
-    BlockIdReader, BlockNumReader, BlockReader, StateProviderBox, StateProviderFactory,
+    BlockIdReader, BlockReader, StateProviderBox, StateProviderFactory,
 };
 use reth_revm::{
     database::StateProviderDatabase,